Revert "Feature auth"
