--- src/js/helpers/initialize-firebase.js
@@ -1,3 +1,0 @@
-import { initializeApp } from 'firebase/app';
-import './firebase-config';
-export default app = initializeApp(firebaseConfig);
