+++ src/partials/home/header.html
@@ -28,7 +28,7 @@
             <a class="reg-btn" name="regBtn" href="#">Registration</a>
           </li>
           <li class="nav-item">
-            <a href="#" class="logout-btn is-hidden" name="logoutBtn">Logout</a>
+            <a class="logout-btn is-hidden" name="logoutBtn" href="#">Logout</a>
           </li>
         </ul>
       </nav>
